+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 * `slug` type fields can now have an empty string or `null` as their `def` value without the string `'none'` populating automatically.
+* The `underline` feature works properly in tiptap toolbar configuration.
 
 ### Changes